Fix charts sizing docs to usable CodeSandbox JS
- remove `ThemeContext` in favor of `useEuiTheme` to allow CodeSandbox usage

+ remove lodash dependency causing codesandbox issues

+ remove unnecessary data.js call - only used once, so keep it in the file

+ misc cleanup
