+++ src-docs/src/views/elastic_charts/data.js
@@ -1,14 +1,3 @@
-export const TIME_DATA_SMALL = [
-  [1551438630000, 8.515625],
-  [1551438660000, 10.796875],
-  [1551438690000, 11.125],
-  [1551438720000, 21.40625],
-  [1551438750000, 17.921875],
-  [1551438780000, 26.640625],
-  [1551438810000, 31.390625],
-  [1551438840000, 23.953125],
-];
-
 export const TIME_DATA = [
   [1551438000000, 8.3203125],
   [1551438030000, 7.9140625],
