+++ src-docs/src/views/elastic_charts/sizes.js
@@ -1,6 +1,5 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import moment from 'moment';
-import { ThemeContext } from '../../components';
 import {
   Chart,
   Settings,
@@ -30,13 +29,17 @@ import {
   EuiCopy,
 } from '../../../../src/components';
 
-import { formatDate, dateFormatAliases } from '../../../../src/services';
+import {
+  formatDate,
+  dateFormatAliases,
+  withEuiTheme,
+} from '../../../../src/services';
 
 import { MultiChartCard, ChartCard } from './shared';
 
 import { TIME_DATA, TIME_DATA_2 } from './data';
 
-export class Sizes extends Component {
+class Sizes extends Component {
   constructor(props) {
     super(props);
 
@@ -169,7 +172,7 @@ export class Sizes extends Component {
       changeDescription,
     } = this.state;
 
-    const isDarkTheme = this.context.theme.includes('dark');
+    const isDarkTheme = this.props.theme.colorMode === 'DARK';
     const theme = isDarkTheme
       ? EUI_CHARTS_THEME_DARK.theme
       : EUI_CHARTS_THEME_LIGHT.theme;
@@ -191,7 +194,7 @@ export class Sizes extends Component {
     }
 
     return (
-      <Fragment>
+      <>
         <EuiPage>
           <EuiPageContent
             role={null}
@@ -369,9 +372,9 @@ export class Sizes extends Component {
             )}
           </EuiCopy>
         </div>
-      </Fragment>
+      </>
     );
   }
 }
 
-Sizes.contextType = ThemeContext;
+export default withEuiTheme(Sizes);
